software/requirements.txt
# Core (works on Mac + Pi)
mido>=1.3.0
python-rtmidi>=1.5.0
flask>=3.0
pyyaml>=6.0

# Hardware MIDI serial (used on Pi for SC16IS752, optional on Mac)
pyserial>=3.5

# Pi-only (install manually on Raspberry Pi)
# RPi.GPIO>=0.7
# smbus2>=0.4
